# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HRLearnium Course Assistant API",
    "description": "Authenticated course Q&A with cited evidence. Natural questions can return a natural grounded explanation (default) or verbatim excerpts (response_mode). Social replies are labelled conversation and contain no course evidence. An LLM backend is required for natural-language answerability and explanations. Configured backend: openai_compatible. ",
    "version": "0.3.0"
  },
  "paths": {
    "/health/live": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Live",
        "operationId": "live_health_live_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Ready",
        "operationId": "ready_health_ready_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/courses/{course_id}/query": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Query",
        "operationId": "query_v1_courses__course_id__query_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/courses/{course_id}/documents": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_v1_courses__course_id__documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_v1_courses__course_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "content"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_v1_courses__course_id__documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponse"
                  },
                  "title": "Response List Documents V1 Courses  Course Id  Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/courses/{course_id}/documents/{document_id}": {
      "put": {
        "tags": [
          "content"
        ],
        "summary": "Replace Document",
        "operationId": "replace_document_v1_courses__course_id__documents__document_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_replace_document_v1_courses__course_id__documents__document_id__put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "content"
        ],
        "summary": "Retire Document",
        "operationId": "retire_document_v1_courses__course_id__documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/courses/{course_id}/excerpts/{excerpt_id}": {
      "get": {
        "tags": [
          "query"
        ],
        "summary": "Get Excerpt",
        "operationId": "get_excerpt_v1_courses__course_id__excerpts__excerpt_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "course_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Course Id"
            }
          },
          {
            "name": "excerpt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Excerpt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Excerpt"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_replace_document_v1_courses__course_id__documents__document_id__put": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_replace_document_v1_courses__course_id__documents__document_id__put"
      },
      "Body_upload_document_v1_courses__course_id__documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_v1_courses__course_id__documents_post"
      },
      "Citation": {
        "properties": {
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "chapter_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chapter Number"
          },
          "chapter_title": {
            "type": "string",
            "title": "Chapter Title"
          },
          "section_kind": {
            "type": "string",
            "title": "Section Kind"
          },
          "paragraph_start": {
            "type": "integer",
            "title": "Paragraph Start"
          },
          "paragraph_end": {
            "type": "integer",
            "title": "Paragraph End"
          },
          "source_start": {
            "type": "integer",
            "title": "Source Start"
          },
          "source_end": {
            "type": "integer",
            "title": "Source End"
          },
          "source_sha256": {
            "type": "string",
            "title": "Source Sha256"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_id",
          "version",
          "document_title",
          "chapter_number",
          "chapter_title",
          "section_kind",
          "paragraph_start",
          "paragraph_end",
          "source_start",
          "source_end",
          "source_sha256"
        ],
        "title": "Citation"
      },
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "passage_count": {
            "type": "integer",
            "title": "Passage Count"
          },
          "paragraph_count": {
            "type": "integer",
            "title": "Paragraph Count"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "filename",
          "version",
          "sha256",
          "active",
          "passage_count",
          "paragraph_count",
          "warnings",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentResponse"
      },
      "Excerpt": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "citation": {
            "$ref": "#/components/schemas/Citation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "text",
          "citation"
        ],
        "title": "Excerpt"
      },
      "Explanation": {
        "properties": {
          "statements": {
            "items": {
              "$ref": "#/components/schemas/GroundedStatement"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Statements"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "statements"
        ],
        "title": "Explanation"
      },
      "GroundedStatement": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 1200,
            "minLength": 1,
            "title": "Text"
          },
          "citation_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 1,
            "title": "Citation Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text",
          "citation_ids"
        ],
        "title": "GroundedStatement"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "QueryRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 2,
            "title": "Question"
          },
          "response_mode": {
            "type": "string",
            "enum": [
              "verbatim",
              "explained"
            ],
            "title": "Response Mode",
            "description": "Both modes accept natural questions. Verbatim returns exact source excerpts; explained (default) returns a natural cited answer; exact evidence remains in excerpts.",
            "default": "explained"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id",
            "description": "Omit this field for your first question. For a follow-up, copy the conversation_id returned by a successful request for this same user and course. Do not invent an ID."
          },
          "include_evaluation": {
            "type": "boolean",
            "title": "Include Evaluation",
            "description": "Include timings, ranking and configuration for evaluation tools.",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question"
        ],
        "title": "QueryRequest",
        "examples": [
          {
            "question": "مدل ۳۰-۳۰-۳۰ چطور به اولویت‌بندی در بحران کمک می‌کند؟",
            "response_mode": "explained"
          }
        ]
      },
      "QueryResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "answered",
              "refused",
              "clarification",
              "conversation"
            ],
            "title": "Status"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "excerpts": {
            "items": {
              "$ref": "#/components/schemas/Excerpt"
            },
            "type": "array",
            "title": "Excerpts"
          },
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "reason_code": {
            "type": "string",
            "title": "Reason Code"
          },
          "retrieval_mode": {
            "type": "string",
            "enum": [
              "lexical",
              "hybrid",
              "hybrid_rerank",
              "full_context"
            ],
            "title": "Retrieval Mode"
          },
          "response_mode": {
            "type": "string",
            "enum": [
              "verbatim",
              "explained"
            ],
            "title": "Response Mode",
            "default": "explained"
          },
          "explanation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Explanation"
              },
              {
                "type": "null"
              }
            ]
          },
          "policy_version": {
            "type": "string",
            "title": "Policy Version",
            "default": "grounded-course-v3"
          },
          "evaluation": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evaluation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "status",
          "answer",
          "excerpts",
          "conversation_id",
          "reason_code",
          "retrieval_mode"
        ],
        "title": "QueryResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}